Add files via upload
Stacking is a technique that associates the prediction of a sub-model with a learning algorithm. It is a very tangle task to be consistent while identifying the rotten fruits manually. Federated learning is a way of training algorithms across multiple decentralized devices, without commuting local data. Deep learning uses neural network simulation to extract features from the provided dataset and make predictions. The image dataset contains noise, which is removed through SMV(Standard Normal Variate) and the image dataset is filtered by a different pre-trained model. Stacking is done on a number of models to get an effective prediction. In this method, the maximum data is used to train the model is 80% while 20% data is used for testing. The up-shot demonstrates that stacking and federated learning approach is more accurate (>95%) and effective.